+++ src/fm_mpx.h
@@ -41,5 +41,5 @@ enum mpx_subcarriers {
 extern void fm_mpx_init(uint32_t sample_rate);
 extern void fm_rds_get_frames(float *outbuf, size_t num_frames);
 extern void fm_mpx_exit();
-extern void set_output_volume(uint8_t vol);
-extern void set_carrier_volume(uint8_t carrier, uint8_t new_volume);
+extern void set_output_volume(float vol);
+extern void set_carrier_volume(uint8_t carrier, float new_volume);
